// src/app/interfaces/create-project-options.ts
export interface CreateProjectOptions {
    title: string;
    company: string;
    description: string;
    project_start: undefined;
    project_end: undefined;
    registration_period: undefined;
    address: string;
    address_comment: string;
    payment: number;
    number_participants: number;
}